refactor: group modern and legacy function implementations
