+++ docs/decisions/2026-09-07-boundary-contract.md
@@ -29,7 +29,7 @@ Zod parsing and encoding are distinct operations. Parsing can apply defaults, fo
 | Client query/mutation/action/subscription | Registry argument schemas encode outgoing values; return schemas parse incoming values. | Missing schemas pass through. Default decode policy warns and returns raw wire data: runtime codec types are **not guaranteed after a decode failure**. `onDecodeError: 'throw'` raises `ZodvexDecodeError` for failed schema parses. Argument encoding errors propagate. Exceptions thrown directly by user codec code can also propagate. |
 | React query hook and watch | Explicit `'skip'` avoids encoding; loading `undefined` passes through. After remediation, failed argument encoding calls the inner hook with skip, then rethrows for React error handling. Watches memoize completed decode results by wire identity. | An encoding failure is not a loading state. Strict decode failures are retried on later reads rather than poisoning the cache. Default warn-mode raw results still follow the configured policy. |
 
-Implementation: [function contracts](../../packages/zodvex/src/internal/functionContracts.ts), [direct wrappers](../../packages/zodvex/src/internal/wrappers.ts), [custom builders](../../packages/zodvex/src/internal/custom.ts), [server errors/returns](../../packages/zodvex/src/internal/serverUtils.ts), [document codecs](../../packages/zodvex/src/internal/codec.ts), [database wrappers](../../packages/zodvex/src/internal/db.ts), [outbound context helpers](../../packages/zodvex/src/internal/actionCtx.ts), [initialization](../../packages/zodvex/src/internal/init.ts), [registry boundary helpers](../../packages/zodvex/src/internal/boundaryHelpers.ts), [React hooks](../../packages/zodvex/src/public/react/hooks.ts), [watch cache](../../packages/zodvex/src/public/react/zodvexReactClient.ts).
+Implementation: [function contracts](../../packages/zodvex/src/internal/functions/contracts.ts), [direct wrappers](../../packages/zodvex/src/internal/legacy/wrappers.ts), [custom builders](../../packages/zodvex/src/internal/functions/customFunctions.ts), [server errors/returns](../../packages/zodvex/src/internal/serverUtils.ts), [document codecs](../../packages/zodvex/src/internal/codec.ts), [database wrappers](../../packages/zodvex/src/internal/db.ts), [outbound context helpers](../../packages/zodvex/src/internal/actionCtx.ts), [initialization](../../packages/zodvex/src/internal/functions/init.ts), [registry boundary helpers](../../packages/zodvex/src/internal/boundaryHelpers.ts), [React hooks](../../packages/zodvex/src/public/react/hooks.ts), [watch cache](../../packages/zodvex/src/public/react/zodvexReactClient.ts).
 
 `skipConvexValidation` on custom builders changes generated Convex validation, not the Zod argument/return work described above. It must not be represented as an equivalent benchmark configuration without naming the changed structural checks. See [its tests](../../packages/zodvex/__tests__/skip-convex-validation.test.ts).
