+++ packages/zodvex/src/internal/functions/init.ts
@@ -12,15 +12,15 @@ import type {
 } from 'convex/server'
 import { NoOp } from 'convex-helpers/server/customFunctions'
 import type { z } from 'zod'
-import { createCodecCallOverrides } from './actionCtx'
-import type { CustomBuilder } from './custom'
-import { zCustomAction, zCustomMutation, zCustomQuery } from './custom'
-import { createZodvexCustomization, type ZodvexUnderlyingDb } from './customization'
-import type { ZodvexDatabaseReader, ZodvexDatabaseWriter } from './db'
-import type { ZodValidator } from './mapping'
-import type { ZodTableMap } from './schema'
-import type { AnyRegistry, Overwrite } from './types'
-import type { $ZodObject } from './zod-core'
+import { createCodecCallOverrides } from '../actionCtx'
+import { createZodvexCustomization, type ZodvexUnderlyingDb } from '../customization'
+import type { ZodvexDatabaseReader, ZodvexDatabaseWriter } from '../db'
+import type { ZodValidator } from '../mapping'
+import type { ZodTableMap } from '../schema'
+import type { AnyRegistry, Overwrite } from '../types'
+import type { $ZodObject } from '../zod-core'
+import type { CustomBuilder } from './customFunctions'
+import { zCustomAction, zCustomMutation, zCustomQuery } from './customFunctions'
 
 /**
  * The context type received by query handlers when wrapDb: true.
