+++ packages/zodvex/src/public/server/index.ts
@@ -14,22 +14,6 @@
 export { customCtx } from 'convex-helpers/server/customFunctions'
 // Action context wrapping (auto-codec for runQuery/runMutation in actions)
 export { createZodvexActionCtx } from '../../internal/actionCtx'
-// Function builders (named — all deprecated)
-export {
-  zActionBuilder,
-  zCustomActionBuilder,
-  zCustomMutationBuilder,
-  zCustomQueryBuilder,
-  zMutationBuilder,
-  zQueryBuilder
-} from '../../internal/builders'
-// Custom function utilities (named — hide customFnBuilder, Overwrite re-export)
-export {
-  type CustomBuilder,
-  zCustomAction,
-  zCustomMutation,
-  zCustomQuery
-} from '../../internal/custom'
 // Codec customization (manual composition escape hatch)
 export {
   createZodvexCustomization,
@@ -54,6 +38,13 @@ export {
   type ZodvexUpperBoundBuilder,
   type ZodvexWriteValue
 } from '../../internal/db'
+// Custom function utilities (named — hide customFnBuilder, Overwrite re-export)
+export {
+  type CustomBuilder,
+  zCustomAction,
+  zCustomMutation,
+  zCustomQuery
+} from '../../internal/functions/customFunctions'
 // One-time setup + types
 export {
   defineContext,
@@ -63,7 +54,16 @@ export {
   type ZodvexCustomization,
   type ZodvexMutationCtx,
   type ZodvexQueryCtx
-} from '../../internal/init'
+} from '../../internal/functions/init'
+// Function builders (named — all deprecated)
+export {
+  zActionBuilder,
+  zCustomActionBuilder,
+  zCustomMutationBuilder,
+  zCustomQueryBuilder,
+  zMutationBuilder,
+  zQueryBuilder
+} from '../../internal/legacy/builders'
 // Rule and audit types for .withRules() and .audit()
 //
 // Rule and audit types (re-exported from ruleTypes.ts via rules.ts)
